Added scheduler

+++ ControlUnit/storage/temperature_threshold.h
@@ -1,13 +1,15 @@
 #ifndef _STORAGE_TEMPERATURE_THRESHOLD_H_
 #define _STORAGE_TEMPERATURE_THRESHOLD_H_
 
+#include <avr/io.h>
+
 /**
  * \brief 
  * Get the current temperature threshold.
  * 
  * \return uint8_t The current temperature threshold
  */
-int16_t get_temperature_threshold(void);
+int8_t get_temperature_threshold(void);
 
 /**
  * \brief 
@@ -15,6 +17,6 @@ int16_t get_temperature_threshold(void);
  * 
  * \param temperature_threshold The temperature threshold to set
  */
-void set_temperature_threshold(int16_t temperature_threshold);
+void set_temperature_threshold(int8_t temperature_threshold);
 
 #endif
